refactor: use langauge in settings
