+++ src/config/settings.ts
@@ -7,13 +7,13 @@ export enum SettingEnum {
 	AI_OPTIONS_ID = 'ai-options',
 	MODEL_ADDRESS_ID = 'self-hosted-model-address-id',
 	OPEN_AI_API_KEY_ID = ' open-ai-api-key-id',
-	OPEN_AI_API_MODEL_ID = 'open-ai-api-model=id',
-	// Mistral_AI_API_KEY_ID = 'mistral-ai-api-key-id',
-	// Mistral_AI_API_MODEL_ID = 'mistral-ai-model-id',
+	OPEN_AI_API_MODEL_ID = 'open-ai-api-model-id',
 	GEMINI_AI_API_KEY_ID = 'gemini-ai-api-key-id',
 	SELF_HOSTED_MODEL = 'self-hosted-model',
 	OPEN_AI = 'open-ai',
 	GEMINI = 'gemini',
+	// Mistral_AI_API_KEY_ID = 'mistral-ai-api-key-id',
+	// Mistral_AI_API_MODEL_ID = 'mistral-ai-model-id',
 	// MISTRAL = 'mistral',
 }
 
@@ -24,15 +24,21 @@ export const settings: Array<ISetting> = [
 		packageValue: SettingEnum.SELF_HOSTED_MODEL,
 		required: true,
 		public: false,
-		i18nLabel: 'Choose AI to use',
-		i18nPlaceholder: 'Choose AI to use',
+		i18nLabel: 'Choose_AI_Model_Label',
+		i18nPlaceholder: 'Choose_AI_Model_Placeholder',
 		values: [
 			{
 				key: SettingEnum.SELF_HOSTED_MODEL,
-				i18nLabel: 'Self Hosted Model',
+				i18nLabel: 'Self_Hosted_AI_Model',
 			},
-			{ key: SettingEnum.OPEN_AI, i18nLabel: 'Open AI' },
-			{ key: SettingEnum.GEMINI, i18nLabel: 'Gemini' },
+			{
+				key: SettingEnum.OPEN_AI,
+				i18nLabel: 'Open_AI',
+			},
+			{
+				key: SettingEnum.GEMINI,
+				i18nLabel: 'Gemini_AI',
+			},
 			// { key: SettingEnum.MISTRAL, i18nLabel: 'Mistral' },
 		],
 	},
@@ -42,8 +48,8 @@ export const settings: Array<ISetting> = [
 		packageValue: 'http://mistral-7b/v1',
 		required: true,
 		public: false,
-		i18nLabel: 'Self Hosted AI Model Address',
-		i18nPlaceholder: 'Self Hosted AI Model Address',
+		i18nLabel: 'Self_Hosted_AI_Model_URL_Label',
+		i18nPlaceholder: 'Self_Hosted_AI_Model_URL_Placeholder',
 	},
 	{
 		id: SettingEnum.OPEN_AI_API_KEY_ID,
@@ -51,8 +57,8 @@ export const settings: Array<ISetting> = [
 		packageValue: '',
 		required: true,
 		public: false,
-		i18nLabel: 'Open AI API Key',
-		i18nPlaceholder: 'Open AI API Key',
+		i18nLabel: 'Open_AI_API_Key_Label',
+		i18nPlaceholder: 'Open_AI_API_Key_Placeholder',
 	},
 	{
 		id: SettingEnum.OPEN_AI_API_MODEL_ID,
@@ -60,8 +66,8 @@ export const settings: Array<ISetting> = [
 		packageValue: '',
 		required: true,
 		public: false,
-		i18nLabel: 'Open AI Model',
-		i18nPlaceholder: 'Open AI API Model',
+		i18nLabel: 'Open_AI_Model',
+		i18nPlaceholder: 'Open_AI_Model_Placeholder',
 	},
 	{
 		id: SettingEnum.GEMINI_AI_API_KEY_ID,
@@ -69,8 +75,8 @@ export const settings: Array<ISetting> = [
 		packageValue: '',
 		required: true,
 		public: false,
-		i18nLabel: 'Gemini AI API Key',
-		i18nPlaceholder: 'Gemini AI API Key',
+		i18nLabel: 'Gemini_API_Key_Label',
+		i18nPlaceholder: 'Gemini_API_Key_Placeholder',
 	},
 	// {
 	// 	id: SettingEnum.Mistral_AI_API_KEY_ID,
